+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+## Enhancements
+* authorisation - added config option to avoid adding a roles column to the users table
+  https://github.com/anvilistas/anvil-extras/pull/516
+
+
 # v2.6.1 27-Mar-2024
 
 ## Bug Fixes
@@ -15,6 +22,8 @@
 ## Minor Changes
 * designer hints - add some designer hints to components
   https://github.com/anvilistas/anvil-extras/pull/512
+* authorisation - added config option to avoid adding a roles column to the users table
+  https://github.com/anvilistas/anvil-extras/pull/516
 
 # v2.6.0 09-Feb-2024